Begin Refactoring Components to Fix Logic in Bookshelf Changes
When options are selected for individual books' target shelves, the entire bookshelf changes to show all books from the *target* bookshelf.

The problem seems to stem from the showBooks component handling too much on its own and combining the tasks of displaying bookshelves, their books, and trying to manage the individual shelf choices for all the books.